pyproject.toml
[tool.poetry]
name = "anycoin"
version = "0.1.0"
description = ""
authors = ["HK-Mattew <mattewsub6565@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/HK-Mattew/anycoin"
repository = "https://github.com/HK-Mattew/anycoin"

packages = [
    { include = "anycoin"}
]


[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.28.1"
pydantic = "^2.10.5"
aiocache = {extras = ["memcached"], version = "^0.12.3"}


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.2"
taskipy = "^1.14.1"
pytest-asyncio = "^0.25.2"
pytest-cov = "^6.0.0"
respx = "^0.22.0"


[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = "session"

[tool.ruff]
line-length = 79

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ["PLR6201"]

[tool.ruff.lint.pylint]
max-args = 10

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
pre_test = 'task lint'
test = 'pytest --cov=anycoin -vv'
post_test = 'coverage html'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
